GNG-418: Refactor Search FilterMenu out of TabMenu component
